EU Constants cleanup
