+++ lib/controllers/european.controller.ts
@@ -1,9 +1,9 @@
 import { Session } from './../../dist/interfaces/common.interfaces.d';
-import { EU_CONSTANTS } from './../constants';
+import { EU_CONSTANTS, EU_BASE_URL } from './../constants/europe';
 import { BlueLinkyConfig } from './../interfaces/common.interfaces';
 import * as pr from 'push-receiver';
 import got from 'got';
-import { ALL_ENDPOINTS, EU_BASE_URL } from '../constants';
+import { ALL_ENDPOINTS } from '../constants';
 import { Vehicle } from '../vehicles/vehicle';
 import EuropeanVehicle from '../vehicles/european.vehicle';
 import { SessionController } from './controller';
